# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/user/CreateUser": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Add New User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Name",
                        "name": "Name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Enter password",
                        "name": "Password",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Enter password again",
                        "name": "rePassword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\", \"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/DeleteUser": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Delete Exist User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deleted ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\", \"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/GetUserList": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "All Users",
                "responses": {
                    "200": {
                        "description": "code\", \"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/UpdateUser": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update User Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "Name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "Password",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\", \"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}